+++ NEWS.md
@@ -2,12 +2,19 @@
 
 - Python background threads can now run in parallel with
   the R session (#1641).
+  
+- Internal updates for NumPy 2.1 (#1651)
 
 - Fixed error when importing a module named `config` (#1628)
 
 - `conda_run2()` is now exported (#1637, contributed by @dramanica)
 
 - Fixes for CRAN check failures (#1645)
+
+- New `as.raw()` method for `python.builtin.bytes` (#1649, #1652)
+
+- `as.character()` method for `python.builtin.bytes` gains a `nul` argument,
+  allowing for convenient handling of embedded NULs in the string. (#1652)
 
 # reticulate 1.38.0